feat(reversinglabs-spectra-intel-submission): add to catalog
