+++ internal-enrichment/reversinglabs-spectra-intel-submission/__metadata__/connector_config_schema.json
@@ -1,0 +1,121 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "https://www.filigran.io/connectors/reversinglabs-spectra-intel-submission_config.schema.json",
+  "type": "object",
+  "properties": {
+    "OPENCTI_URL": {
+      "description": "The base URL of the OpenCTI instance.",
+      "format": "uri",
+      "maxLength": 2083,
+      "minLength": 1,
+      "type": "string"
+    },
+    "OPENCTI_TOKEN": {
+      "description": "The API token to connect to OpenCTI.",
+      "type": "string"
+    },
+    "CONNECTOR_NAME": {
+      "default": "ReversingLabs Spectra Intelligence Submission",
+      "description": "The name of the connector.",
+      "type": "string"
+    },
+    "CONNECTOR_SCOPE": {
+      "default": [
+        "Artifact",
+        "Url",
+        "StixFile",
+        "File"
+      ],
+      "description": "The scope of the connector.",
+      "items": {
+        "type": "string"
+      },
+      "type": "array"
+    },
+    "CONNECTOR_LOG_LEVEL": {
+      "default": "error",
+      "description": "The minimum level of logs to display.",
+      "enum": [
+        "debug",
+        "info",
+        "warn",
+        "warning",
+        "error"
+      ],
+      "type": "string"
+    },
+    "CONNECTOR_TYPE": {
+      "const": "INTERNAL_ENRICHMENT",
+      "default": "INTERNAL_ENRICHMENT",
+      "type": "string"
+    },
+    "CONNECTOR_AUTO": {
+      "default": false,
+      "description": "Whether the connector should run automatically when an entity is created or updated.",
+      "type": "boolean"
+    },
+    "REVERSINGLABS_SPECTRA_INTEL_SUBMISSION_URL": {
+      "default": "data.reversinglabs.com",
+      "description": "ReversingLabs Spectra Intelligence API base URL.",
+      "type": "string"
+    },
+    "REVERSINGLABS_SPECTRA_INTEL_SUBMISSION_USERNAME": {
+      "description": "ReversingLabs Spectra Intelligence username.",
+      "type": "string"
+    },
+    "REVERSINGLABS_SPECTRA_INTEL_SUBMISSION_PASSWORD": {
+      "description": "ReversingLabs Spectra Intelligence password.",
+      "format": "password",
+      "type": "string",
+      "writeOnly": true
+    },
+    "REVERSINGLABS_SPECTRA_INTEL_SUBMISSION_MAX_TLP": {
+      "default": "TLP:AMBER",
+      "description": "Maximum TLP level for entities that the connector can enrich.",
+      "enum": [
+        "TLP:WHITE",
+        "TLP:CLEAR",
+        "TLP:GREEN",
+        "TLP:AMBER",
+        "TLP:AMBER+STRICT",
+        "TLP:RED"
+      ],
+      "type": "string"
+    },
+    "REVERSINGLABS_SPECTRA_INTEL_SUBMISSION_SANDBOX_OS": {
+      "default": "windows10",
+      "description": "The platform to execute the sample on.",
+      "enum": [
+        "windows7",
+        "windows10",
+        "windows11",
+        "macos11",
+        "linux"
+      ],
+      "type": "string"
+    },
+    "REVERSINGLABS_SPECTRA_INTEL_SUBMISSION_SANDBOX_INTERNET_SIM": {
+      "default": false,
+      "description": "Enable internet simulation during sandbox analysis.",
+      "type": "boolean"
+    },
+    "REVERSINGLABS_SPECTRA_INTEL_SUBMISSION_CREATE_INDICATORS": {
+      "default": true,
+      "description": "Create STIX indicators from analysis results.",
+      "type": "boolean"
+    },
+    "REVERSINGLABS_SPECTRA_INTEL_SUBMISSION_POLL_INTERVAL": {
+      "default": 250,
+      "description": "Polling interval in seconds to check analysis results.",
+      "minimum": 250,
+      "type": "integer"
+    }
+  },
+  "required": [
+    "OPENCTI_URL",
+    "OPENCTI_TOKEN",
+    "REVERSINGLABS_SPECTRA_INTEL_SUBMISSION_USERNAME",
+    "REVERSINGLABS_SPECTRA_INTEL_SUBMISSION_PASSWORD"
+  ],
+  "additionalProperties": true
+}
