+++ internal-enrichment/reversinglabs-spectra-intel-submission/__metadata__/connector_manifest.json
@@ -14,7 +14,7 @@
   "support_version": ">=6.2.14",
   "subscription_link": null,
   "source_code": "https://github.com/OpenCTI-Platform/connectors/tree/master/internal-enrichment/reversinglabs-spectra-intel-submission",
-  "manager_supported": false,
+  "manager_supported": true,
   "container_version": "rolling",
   "container_image": "opencti/connector-reversinglabs-spectra-intel-submission",
   "container_type": "INTERNAL_ENRICHMENT"
