+++ internal-enrichment/reversinglabs-spectra-intel-submission/entrypoint.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 # Correct working directory
-cd /opt/opencti-connector-reversinglabs-spectra-intelligence-file-submission
+cd /opt/opencti-connector-reversinglabs-spectra-intel-submission
 
 # Launch the connector
 python3 main.py
