+++ internal-enrichment/reversinglabs-spectra-intel-submission/requirements.txt
@@ -1,2 +1,4 @@
 pycti==7.260626.0
 reversinglabs-sdk-py3==2.11.4
+pydantic>=2.8.2,<3
+connectors-sdk @ git+https://github.com/OpenCTI-Platform/connectors.git@master#subdirectory=connectors-sdk
