+++ internal-enrichment/reversinglabs-spectra-intel-submission/src/__init__.py
@@ -1,0 +1,3 @@
+from connector.settings import ConnectorSettings
+
+__all__ = ["ConnectorSettings"]
