+++ internal-enrichment/reversinglabs-spectra-intel-submission/src/connector/__init__.py
@@ -1,0 +1,4 @@
+from connector.connector import ReversingLabsSpectraIntelConnector
+from connector.settings import ConnectorSettings
+
+__all__ = ["ConnectorSettings", "ReversingLabsSpectraIntelConnector"]
