+++ internal-enrichment/reversinglabs-spectra-intel-submission/tests/test-requirements.txt
@@ -1,0 +1,3 @@
+# Main dependencies needs to be installed
+-r ../requirements.txt
+pytest==8.4.2
